==> inspi-score/3_t_label_v1.ipynb <==
import pandas as pd

import spacy
import collections

from bs4 import BeautifulSoup as BS

nlp = spacy.load('de_core_news_sm')

pd.set_option('display.max_rows', 500)
pd.set_option('display.max_columns', 500)
pd.set_option('display.width', 1000)

df_tag_category = pd.read_parquet('data/tag_category.parquet')
df_tag = pd.read_parquet('data/tag.parquet')

df_event = pd.read_parquet('data/event.parquet')
df_event_tags = pd.read_parquet('data/event_tags.parquet')
df_event_material_list = pd.read_parquet('data/event_material_list.parquet')

df_experiment = pd.read_parquet('data/experiment.parquet')
df_experiment_item = pd.read_parquet('data/experiment_item.parquet')

df_feature_vector = pd.read_parquet('data/feature_vector.parquet')

df_feature_vector = df_feature_vector[[
  'id',
  'title',
  'description',

  'p_tag',
  'img_tag',
  'b_tag',
  'li_tag',
  'header_img',
  'text_len',
  'span_tag',
  'a_tag',
  'u_tag',

  'costsRating',
  'executionTimeRating',
  'isPrepairationNeeded',

  'total_words',

  'nomen_quote',
  'adjektive_quote',
  'numerisch_quote',
  'satzzeichen_quote',
  'konjunktion_quote',

  'count_question_mark'
]]

def percentage_change(col1, col2):
    return col1 / col2 * 100

def calc_label_score(unklar, doof, gut, sehr_gut, count):
    score = (doof * 1 + gut * 3 + sehr_gut * 5) / count
    return score

df_label_temp = df_experiment_item.groupby(["event_id_id"]).size().reset_index(name='Bewertungen')
df_1 = df_experiment_item[df_experiment_item['score'] == 1].groupby(["event_id_id"]).size().reset_index(name='doof')
df_2 = df_experiment_item[df_experiment_item['score'] == 2].groupby(["event_id_id"]).size().reset_index(name='gut')
df_3 = df_experiment_item[df_experiment_item['score'] == 3].groupby(["event_id_id"]).size().reset_index(name='sehr_gut')
df_4 = df_experiment_item[df_experiment_item['score'] == 4].groupby(["event_id_id"]).size().reset_index(name='unklar')

df_label_temp_2 = df_label_temp.join(
    df_1.set_index('event_id_id'), on='event_id_id').join(
    df_2.set_index('event_id_id'), on='event_id_id').join(
    df_3.set_index('event_id_id'), on='event_id_id').join(
    df_4.set_index('event_id_id'), on='event_id_id')

df_label_temp_2 = df_label_temp_2.fillna(0)

df_label = pd.merge(df_event, df_label_temp_2, left_on=['id'], right_on= ['event_id_id'], how = 'left')

df_label['doof_rate'] = percentage_change(df_label['doof'], df_label['Bewertungen'])
df_label['gut_rate'] = percentage_change(df_label['gut'], df_label['Bewertungen'])
df_label['sehr_gut_rate'] = percentage_change(df_label['sehr_gut'], df_label['Bewertungen'])
df_label['unklar_rate'] = percentage_change(df_label['unklar'], df_label['Bewertungen'])

df_label['calc_label_score'] = calc_label_score(df_label['unklar'], df_label['doof'], df_label['gut'], df_label['sehr_gut'], df_label['Bewertungen'])

# df = pd.merge(df_label, event_tags_features, left_on=['id'], right_on= ['event_id'], how = 'left')
# df_raw = df

df_label = df_label[df_label['Bewertungen'] > 1]
df_label = df_label[df_label['unklar'] == 0]

df_label = df_label.round(2)

df_label.to_parquet('data/label_v1.parquet', coerce_timestamps="us")

df_label.nlargest(5, 'unklar_rate')

df_label.nlargest(5, 'gut_rate')

df_label.nlargest(5, 'sehr_gut_rate')

df_label.nlargest(5, 'doof_rate')
